requirement.txt
asgiref==3.11.0
Django==5.2.10
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
gunicorn==25.0.0
PyMySQL==1.1.2
packaging==26.0
Pillow==12.1.0
PyJWT==2.11.0
python-dotenv==1.2.1
sqlparse==0.5.5
typing_extensions==4.15.0